Rework commit planning to be simpler and more robust
Unlike Test::Chimps, from which some parts of this project are
derived, the scheduler is able to store a set of future tests to plan.
Taking advantage of this by simply adding to Gearman's queue every
commit between the old ref and the new one, we can remove the need to
"TestedHead"s, and much of the complicated Project->schedule_tests
logic.