Only test commits which _no_ branch has covered
If a merge pulls in a branch whose tip is tested, but which also pulls
in earlier, untested commits, trust the tip's status, and skip testing
all of those earlier untested commits which it also pulls in.  We do
this by excluding _all_ tested commit shas, not just the current
branch's.