Move to Jifty::Plugin::RPC instead of Gearman
Gearman, while it has its uses, does not have guaranteed execution.
Make use of the Jifty::Plugin::RPC plugin to use AMQP for a back-end
message bus for RPC; this provides guaranteed execution as well as a
generalized message bus for other uses.  Guaranteed execution obviates
the need for a check_queue call or method, but does require that the
`gearman_process` column be renamed to the more general `queue_status`
column.

This commit removes the real-time completion updates that
$status->percent provided; such functionality will be later replaced by
the message bus.