Never assign to global (non-lexical) $_, due to AnyEvent event loop
Any place wherein a postfix while assigns to $_, it is modifying the
global $_, not the lexical $_ that for and map make use of.  This leads
to the possibility of action at a distance -- particularly when in an
event-driven context such as AnyEvent.

Observed failure modes of global $_ being modified when running under
Twiggy include unpredictable server hangs and spurious errors; these
were much more common under 5.16.0 than under 5.10.1, for whatever
reason.